My Plant Friends Interactive Teaching System/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
python-dotenv==1.0.0
pydantic==2.3.0
pydantic-settings==2.0.3
motor==3.3.1
websockets==11.0.3
python-multipart